// src/API/NauthApi_gen/models/AuthFailureReasons.ts
/* tslint:disable */
/* eslint-disable */
/**
 * nauth-asp
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 */
export const AuthFailureReasons = {
  SessionExpired: "SessionExpired",
  _2FaRequired: "_2FARequired",
  RequireVerifiedEmail: "RequireVerifiedEmail",
  RequireEnabledUser: "RequireEnabledUser",
  Require2FaSetup: "Require2FASetup",
  ForeginResource: "ForeginResource",
  PrManageUsers: "PrManageUsers",
  PrManageOwnServices: "PrManageOwnServices",
  PrManageServices: "PrManageServices",
  PrManageEmailTemplates: "PrManageEmailTemplates",
} as const;
export type AuthFailureReasons = (typeof AuthFailureReasons)[keyof typeof AuthFailureReasons];

export function instanceOfAuthFailureReasons(value: any): boolean {
  for (const key in AuthFailureReasons) {
    if (Object.prototype.hasOwnProperty.call(AuthFailureReasons, key)) {
      if (AuthFailureReasons[key as keyof typeof AuthFailureReasons] === value) {
        return true;
      }
    }
  }
  return false;
}

export function AuthFailureReasonsFromJSON(json: any): AuthFailureReasons {
  return AuthFailureReasonsFromJSONTyped(json, false);
}

export function AuthFailureReasonsFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthFailureReasons {
  return json as AuthFailureReasons;
}

export function AuthFailureReasonsToJSON(value?: AuthFailureReasons | null): any {
  return value as any;
}

export function AuthFailureReasonsToJSONTyped(value: any, ignoreDiscriminator: boolean): AuthFailureReasons {
  return value as AuthFailureReasons;
}