// src/API/NauthApi_gen/models/StringResponseWrapper.ts
/* tslint:disable */
/* eslint-disable */
/**
 * nauth-asp
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
import type { WrResponseStatus } from "./WrResponseStatus";
import { WrResponseStatusFromJSON, WrResponseStatusFromJSONTyped, WrResponseStatusToJSON, WrResponseStatusToJSONTyped } from "./WrResponseStatus";
import type { AuthFailureReasons } from "./AuthFailureReasons";
import {
  AuthFailureReasonsFromJSON,
  AuthFailureReasonsFromJSONTyped,
  AuthFailureReasonsToJSON,
  AuthFailureReasonsToJSONTyped,
} from "./AuthFailureReasons";

/**
 *
 * @export
 * @interface StringResponseWrapper
 */
export interface StringResponseWrapper {
  /**
   *
   * @type {WrResponseStatus}
   * @memberof StringResponseWrapper
   */
  status?: WrResponseStatus;
  /**
   *
   * @type {Array<AuthFailureReasons>}
   * @memberof StringResponseWrapper
   */
  authenticationFailureReasons?: Array<AuthFailureReasons> | null;
  /**
   *
   * @type {string}
   * @memberof StringResponseWrapper
   */
  response?: string | null;
}

/**
 * Check if a given object implements the StringResponseWrapper interface.
 */
export function instanceOfStringResponseWrapper(value: object): value is StringResponseWrapper {
  return true;
}

export function StringResponseWrapperFromJSON(json: any): StringResponseWrapper {
  return StringResponseWrapperFromJSONTyped(json, false);
}

export function StringResponseWrapperFromJSONTyped(json: any, ignoreDiscriminator: boolean): StringResponseWrapper {
  if (json == null) {
    return json;
  }
  return {
    status: json["status"] == null ? undefined : WrResponseStatusFromJSON(json["status"]),
    authenticationFailureReasons:
      json["authenticationFailureReasons"] == null ? undefined : (json["authenticationFailureReasons"] as Array<any>).map(AuthFailureReasonsFromJSON),
    response: json["response"] == null ? undefined : json["response"],
  };
}

export function StringResponseWrapperToJSON(json: any): StringResponseWrapper {
  return StringResponseWrapperToJSONTyped(json, false);
}

export function StringResponseWrapperToJSONTyped(value?: StringResponseWrapper | null, ignoreDiscriminator: boolean = false): any {
  if (value == null) {
    return value;
  }

  return {
    status: WrResponseStatusToJSON(value["status"]),
    authenticationFailureReasons:
      value["authenticationFailureReasons"] == null ? undefined : (value["authenticationFailureReasons"] as Array<any>).map(AuthFailureReasonsToJSON),
    response: value["response"],
  };
}